requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
plotly>=5.18.0
nltk>=3.8.0
bertopic>=0.15.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
sentence-transformers>=2.2.0
pyarrow>=14.0.0
tabulate>=0.9.0
chromadb>=0.4.0

# Phase 7.3: Multi-Source Integration Dependencies
aiohttp>=3.9.0
feedparser>=6.0.10
python-dateutil>=2.8.2
tweepy>=4.14.0
pydantic>=2.0.0

# Phase 8: Executive Briefing PDF Generation Dependencies
fpdf2>=2.7.0
reportlab>=4.0.0
Pillow>=10.0.0
openai>=1.0.0

# Phase 9: Advanced Analytics & Intelligence Dependencies
scipy>=1.11.0
statsmodels>=0.14.0
seaborn>=0.13.0
networkx>=3.2.1
wordcloud>=1.9.0
textblob>=0.17.1
spacy>=3.7.0
gensim>=4.3.0
pyLDAvis>=3.4.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0
yfinance>=0.2.0
alpha-vantage>=2.3.0
feedparser>=6.0.10
xgboost>=2.0.0
lightgbm>=4.0.0

# Phase 10: Automation & Deployment Dependencies
celery>=5.3.0
redis>=5.0.0
boto3>=1.29.0
psycopg2-binary>=2.9.0
cryptography>=41.0.0
gunicorn>=21.0.0
nginx-configparser>=1.0.0

# Portfolio Enhancement Dependencies
hypothesis>=6.92.0
pytest-cov>=4.1.0
